Support tailscale from distfeeds

+++ openwrt/tailscale/shell.dispose.sh
@@ -12,10 +12,17 @@ tailscale down
 /etc/init.d/tailscale disable
 ls /etc/rc.d/S*tailscale*
 
-kill $(ps | grep '[p]/etc/config/tailscaled.state' | awk '{print $1}')
+from_distfeeds=$(opkg list-installed | grep tailscale)
 
-rm /etc/config/tailscaled.state
-rm /etc/init.d/tailscale
-rm /usr/bin/tailscale
-rm /usr/bin/tailscaled
-rm /tmp/tailscale.up.log
+if [ -z "$from_distfeeds" ]
+then
+    kill $(ps | grep '[p]/etc/config/tailscaled.state' | awk '{print $1}')
+
+    rm /etc/config/tailscaled.state
+    rm /etc/init.d/tailscale
+    rm /usr/bin/tailscale
+    rm /usr/bin/tailscaled
+    rm /tmp/tailscale.up.log
+else
+    opkg uninstall tailscale
+fi
